Complete(Registration, Login, Logout, Forgot Password)
Registration page, Registration servlet registration helper class include all the criteria that makes registration clean. Login page , Login servlet with the helping class that also include some criteria for login. homepage, logout page, logout code, session create, forgot password is carried out by the packages forgotpassword, codematching and updatepassword.
It have all the dependencies like mysql-java connector, javax.mail, and many more dependencies.
path are accurately given in web.xml file.
Code matching is done with the help of mail and database.
Some random code is generated with the help of Random class such that it is simultaneously saved in the table resetpasswordrecord of the database coursework and code is sent via ail to the user.
The registration record of the users are stored in users table of the database coursework.
Dependencies can be easily added in the file pom.xml inside project file by searching it from the mvnrepository on internet.