New update for Admin registration
When someone try to register as an admin then the system email a code to the random admin(among we 5) of our system.
then he/she trying to register as admin  may contact with the admins to get code so that admins will get chance to verify that person in clear.
and hence on entering genuine gmail account and code he/she will be the the admin of our system.